requirements.txt
pandas
numpy
xgboost
scikit-learn
plotly
pyyaml
python-dotenv
groq
entsoe-py
joblib
requests